refactor(rtorrent): extract XML-RPC codec into xmlrpc.rs
services/download_client/rtorrent/mod.rs was 2104 LoC; the bottom-of-
file XML-RPC codec was the largest self-contained unit at ~360 LoC.
Moves to a sibling so the trait impl and the wire codec stay
independently navigable:

  mod.rs    1746  (was 2104) — RtorrentClient, the DownloadClient
                   trait impl, and all XML-RPC tests + the
                   send/wait_metadata helpers. Adds `mod xmlrpc;` plus
                   `use xmlrpc::{XmlValue, decode_response,
                   encode_request, xml_attr_escape};` at the top so the
                   trait impl + tests stay terse.
  xmlrpc.rs  370   The XmlValue enum, the encoder (encode_request,
                   encode_value, xml_text_escape, xml_attr_escape),
                   the decoder (decode_response, fault_message,
                   xml_text_unescape, decode_value), and the small
                   pull-Parser. Items used from mod.rs are pub(super);
                   the rest stay private.

The XML-RPC tests (encode_request_*, encode_escapes_*, xml_attr_escape_*,
decode_*, parser_*, fault_message_*) stay inline in mod.rs's `mod tests`
since they read alongside the parent's other unit tests
(endpoint_*, content_path_*, state_mapping_*) and `use super::*;` was
already pulling the codec items in via `use xmlrpc::*` style — the
new top-of-file imports surface them under the same names.

Verify: 1233 unit + 54 rtorrent + 7 PR0 + 1 e2e tests all green,
fmt + clippy -D warnings clean.
